adding gradle functionality
