+++ Search/BingVisualSearch/.gitignore
@@ -1,50 +1,5 @@
-*.class
+# Ignore Gradle project-specific cache directory
+.gradle
 
-# Auth filed
-*.auth
-*.azureauth
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-
-# Azure Tooling #
-node_modules
-packages
-
-# Eclipse #
-*.pydevproject
-.project
-.metadata
-bin/**
-tmp/**
-tmp/**/*
-*.tmp
-*.bak
-*.swp
-*~.nib
-local.properties
-.classpath
-.settings/
-.loadpath
-
-# Other Tooling #
-.classpath
-.project
-target/
-.idea
-*.iml
-
-# Mac OS #
-.DS_Store
-.DS_Store?
-
-# Windows #
-Thumbs.db
-
-# reduced pom files should not be included
-dependency-reduced-pom.xml
+# Ignore Gradle build output directory
+build
