+++ Search/BingVisualSearch/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    java
+    application
+}
+application {
+    mainClassName = "main.java.BingVisualSearchSample"
+}
+repositories {
+    mavenCentral()
+}
+dependencies {
+    compile("org.slf4j:slf4j-simple:1.7.25")
+    compile("com.microsoft.azure.cognitiveservices:azure-cognitiveservices-visualsearch:1.0.2-beta")
+    compile("com.google.code.gson:gson:2.8.5")
+}
