+++ Search/BingVisualSearch/gradle/config/settings
@@ -1,0 +1,116 @@
+# name: Pressing Ctrl-D exits session
+# type: bool
+# Ctrl-D exits cmd.exe when it is pressed on an empty line.
+ctrld_exits = 1
+
+# name: Toggle if pressing Esc clears line
+# type: bool
+# Clink clears the current line when Esc is pressed (unless Readline's Vi mode
+# is enabled).
+esc_clears_line = 1
+
+# name: Match display colour
+# type: int
+# Colour to use when displaying matches. A value less than 0 will be the
+# opposite brightness of the default colour.
+match_colour = -1
+
+# name: Executable match style
+# type: enum
+#  0 = PATH only
+#  1 = PATH and CWD
+#  2 = PATH, CWD, and directories
+# Changes how Clink will match executables when there is no path separator on
+# the line. 0 = PATH only, 1 = PATH and CWD, 2 = PATH, CWD, and directories. In
+# all cases both executables and directories are matched when there is a path
+# separator present. A value of -1 will disable executable matching completely.
+exec_match_style = 2
+
+# name: Whitespace prefix matches files
+# type: bool
+# If the line begins with whitespace then Clink bypasses executable matching and
+# will match all files and directories instead.
+space_prefix_match_files = 1
+
+# name: Colour of the prompt
+# type: int
+# Surrounds the prompt in ANSI escape codes to set the prompt's colour. Disabled
+# when the value is less than 0.
+prompt_colour = -1
+
+# name: Auto-answer terminate prompt
+# type: enum
+#  0 = Disabled
+#  1 = Answer 'Y'
+#  2 = Answer 'N'
+# Automatically answers cmd.exe's 'Terminate batch job (Y/N)?' prompts. 0 =
+# disabled, 1 = answer 'Y', 2 = answer 'N'.
+terminate_autoanswer = 0
+
+# name: Lines of history saved to disk
+# type: int
+# When set to a positive integer this is the number of lines of history that
+# will persist when Clink saves the command history to disk. Use 0 for infinite
+# lines and <0 to disable history persistence.
+history_file_lines = 10000
+
+# name: Skip adding lines prefixed with whitespace
+# type: bool
+# Ignore lines that begin with whitespace when adding lines in to the history.
+history_ignore_space = 0
+
+# name: Controls how duplicate entries are handled
+# type: enum
+#  0 = Always add
+#  1 = Ignore
+#  2 = Erase previous
+# If a line is a duplicate of an existing history entry Clink will erase the
+# duplicate when this is set 2. A value of 1 will not add duplicates to the
+# history and a value of 0 will always add lines. Note that history is not
+# deduplicated when reading/writing to disk.
+history_dupe_mode = 2
+
+# name: Read/write history file each line edited
+# type: bool
+# When non-zero the history will be read from disk before editing a new line and
+# written to disk afterwards.
+history_io = 0
+
+# name: Sets how command history expansion is applied
+# type: enum
+#  0 = Off
+#  1 = On
+#  2 = Not in single quotes
+#  3 = Not in double quote
+#  4 = Not in any quotes
+# The '!' character in an entered line can be interpreted to introduce words
+# from the history. This can be enabled and disable by setting this value to 1
+# or 0. Values or 2, 3 or 4 will skip any ! character quoted in single, double,
+# or both quotes respectively.
+history_expand_mode = 4
+
+# name: Support Windows' Ctrl-Alt substitute for AltGr
+# type: bool
+# Windows provides Ctrl-Alt as a substitute for AltGr, historically to support
+# keyboards with no AltGr key. This may collide with some of Readline's
+# bindings.
+use_altgr_substitute = 1
+
+# name: Strips CR and LF chars on paste
+# type: enum
+#  0 = Paste unchanged
+#  1 = Strip
+#  2 = As space
+# Setting this to a value >0 will make Clink strip CR and LF characters from
+# text pasted into the current line. Set this to 1 to strip all newline
+# characters and 2 to replace them with a space.
+strip_crlf_on_paste = 2
+
+# name: Enables basic ANSI escape code support
+# type: bool
+# When printing the prompt, Clink has basic built-in support for SGR ANSI escape
+# codes to control the text colours. This is automatically disabled if a third
+# party tool is detected that also provides this facility. It can also be
+# disabled by setting this to 0.
+ansi_code_support = 1
+
